Forgive me father for I have sinned and created a jumbo commit, purely out of fear of losing all the work.

+++ src/core/types.h
@@ -15,7 +15,7 @@ typedef struct {
   int channels;
 } ImageMeta;
 
-typedef uint32_t AssetId;
+typedef size_t AssetId;
 typedef uint32_t SignatureT;
 typedef int EntityId;
 typedef size_t EntityIndex;
